snap: move fstype detection for squashfs into osutil/squashfs
