+++ osutil/squashfs/fstype.go
