+++ systemd/systemd.go
@@ -34,7 +34,7 @@ import (
 
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/osutil"
-	"github.com/snapcore/snapd/snap/squashfs"
+	"github.com/snapcore/snapd/osutil/squashfs"
 )
 
 var (
